LayoutTests/accessibility/ios-simulator/strong-password-field-expected.txt


This tests the strong password fields on iOS

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axPw1.boolAttributeValue('AXIsStrongPasswordField') is false
PASS axPw2.boolAttributeValue('AXIsStrongPasswordField') is true
PASS axPw1.stringValue is 'AXValue: ••••••••'
PASS axPw2.stringValue is 'AXValue: strong password'
PASS successfullyParsed is true

TEST COMPLETE
